refactoring options validation test to prevent needless duplication
