+++ lib/OptionsValidationError.js
@@ -2,7 +2,7 @@
 
 /* eslint no-param-reassign: 'off' */
 
-const optionsSchema = require('./optionsSchema.json');
+const optionsSchema = require('./schemas/options.json');
 
 const indent = (str, prefix, firstLine) => {
   if (firstLine) {
