+++ lib/Server.js
@@ -20,7 +20,7 @@ const spdy = require('spdy');
 const webpack = require('webpack');
 const webpackDevMiddleware = require('webpack-dev-middleware');
 const OptionsValidationError = require('./OptionsValidationError');
-const optionsSchema = require('./optionsSchema.json');
+const optionsSchema = require('./schemas/options.json');
 
 require('./polyfills');
